requirements_pip.txt
acoular==23.6
altair==5.0.1
appnope==0.1.2
asttokens==2.0.5
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
blosc2==2.0.0
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
comm==0.1.2
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.4
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.0.4
executing==0.8.3
fonttools==4.25.0
gitdb==4.0.7
GitPython==3.1.37
h5py==3.10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==6.1.0
ipykernel==6.25.0
ipython==8.15.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
libusb==1.0.26b5
llvmlite==0.41.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.0
msgpack==1.0.7
munkres==1.1.4
nest-asyncio==1.5.6
numba==0.58.1
numexpr==2.8.7
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.3
pkg-about==1.0.8
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.15.1
Pympler==0.9
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyusb==1.2.1
pyzmq==25.1.0
requests==2.31.0
rich==13.3.5
scikit-learn==1.1.3
scipy==1.11.3
setuptools==68.0.0
sip==6.6.2
six==1.16.0
smmap==4.0.0
sounddevice==0.4.6
stack-data==0.2.0
streamlit==1.24.1
tables==3.8.0
tenacity==8.2.2
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.7.1
traits==6.4.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==2.1
urllib3==1.26.16
validators==0.18.2
wcwidth==0.2.5
wheel==0.41.2
widgetsnbextension==4.0.5
zipp==3.11.0
